refactor: 자동차 경주의 우승자를 찾고 출력하는 책임을 CarWinner에게 할당
